Remove unused code & move files to proper package
