+++ sqlYourCsv/src/main/java/com/codecool/sqlyourcsv/dataProvider/GoogleAuthorizeUtil.java
@@ -1,4 +1,4 @@
-package com.codecool.sheetsTest;
+package com.codecool.sqlyourcsv.dataProvider;
 
 import com.google.api.client.auth.oauth2.Credential;
 import com.google.api.client.extensions.java6.auth.oauth2.AuthorizationCodeInstalledApp;
